+++ src/app/page.tsx
@@ -1,15 +1,13 @@
 import type { Metadata } from 'next'
 import { Title } from '@/components/Title'
-import { Table } from '@/components/Index/Table'
 import { Flex } from '@/components/Flex'
 import Image from 'next/legacy/image'
 import { DEFAULT_COVER, MAILTO, TEL, TELEPHONE, YANDEX_LOCATION } from '@/constants'
 import { TextBlock } from '@/components/TextBlock'
 import { Partners } from '@/components/Partners'
 import Carousel from '@/components/Carousel'
-import { WideBlock } from '@/components/WideBlock'
 import { Hero } from '@/components/Index/Hero'
-import { TableButtons } from '@/components/Index/TableButtons'
+import { IndexTables } from '@/components/Index/IndexTables'
 import Link from 'next/link'
 import s from '@/index.module.css'
 
@@ -22,581 +20,6 @@ export const metadata: Metadata = {
         ],
     },
 }
-
-const tableDataMp = [
-    {
-        name: 'Рамочная концепция жилого района "Донские Легенды" и детализированные концепции благоустройства 1-й очереди строительства',
-        href: 'https://samolet.ru/rostov-na-donu/project/donskie-legendy/',
-        client: 'ГК «Самолет»',
-        partners: 'Консультанты: ООО Лабград, ОрганикПанк',
-        year: '2024',
-        location: 'Ростов-на-Дону, Ростовская область',
-    },
-    {
-        name: 'Мастерплан ГО Горно-Алтайск (Концепция пространственного развития)',
-        href: '',
-        client: 'ГАУ «Институт Генплана Москвы», АО «ДОМ.РФ»',
-        partners: '',
-        year: '2024',
-        location: 'Горно-Алтайск, Республика Алтай',
-    },
-    {
-        name: 'Мастерплан агломерации Тикси-Найба',
-        href: '',
-        client: 'АО «ДОМ.РФ», ИТП «Урбаника»',
-        partners: '',
-        year: '2024',
-        location: 'Тикси, Республика Саха (Якутия)',
-    },
-    {
-        name: 'Архитектурно-градостроительная концепция комплексного развития территории',
-        href: '',
-        client: 'ООО "СЗ "СПМК-8", АО «ДОМ.РФ»',
-        partners: '',
-        year: '2023',
-        location: 'Чебоксары, Чувашская республика',
-    },
-    {
-        name: 'Дизайн-код города Комсомольск-на-Амуре в рамках мастер-плана',
-        href: '',
-        client: 'ИТП «Урбаника»',
-        partners: '',
-        year: '2023',
-        location: 'Комсомольск-на-Амуре, Хабаровский край',
-    },
-    {
-        name: 'Мастерплан агломерации Кавказских Минеральных Вод (Концепция пространственного развития)',
-        href: '',
-        client: 'ГАУ «Институт Генплана Москвы», АО «ДОМ.РФ»',
-        partners: 'ООО ИТП «Урбаника»',
-        year: '2022',
-        location: 'Кавказские минеральные воды, Ставропольский край',
-    },
-    {
-        name: 'Первое место в открытом всероссийском конкурсе по созданию мастер-плана г.Якутска',
-        href: '/yakutsk',
-        client: 'Агентство стратегического развития «ЦЕНТР», НО "Целевой Фонд будущих поколений Республики Саха"',
-        partners: 'ГАУ «Институт Генплана Москвы», ООО ИТП «Урбаника», АО «Найт Фрэнк», ИП Бейлин М.А., ООО «ЛСТК-Проект», ИТИ СВФУ им. М.К. Аммосова',
-        year: '2022',
-        location: 'Якутск, Республика Саха - Якутия',
-    },
-    {
-        name: 'Эскизный проект пользовательского контура, генеральный план, благоустройство, архитектурные решения и интерьеры Завода ЕАС Нарофоминск',
-        href: '',
-        client: 'ООО "Коттех"',
-        partners: '',
-        year: '2022',
-        location: 'Нарофоминск, Московская область',
-    },
-    {
-        name: 'Мастерплан Волхова',
-        href: '/volkhov',
-        client: 'Администрация Волховского района',
-        partners: 'ООО ИТП «Урбаника»',
-        year: '2021',
-        location: 'Волхов, Ленинградская область',
-    },
-    {
-        name: 'Второе место в конкурсе на мастерплан Экополиса и разработка WEB-3д модели нового города',
-        href: '',
-        client: 'Агентство стратегического развития «ЦЕНТР»',
-        partners: 'Новая Земля',
-        year: '2021',
-        location: 'Корсаков, Сахалинская область',
-    },
-    {
-        name: 'Организация и модерация воркшопа по ревитализации территории завода',
-        href: '',
-        client: 'ООО "Охта Групп"',
-        partners: '',
-        year: '2021',
-        location: 'Санкт-Петербург',
-    },
-    {
-        name: 'Алгоритмический анализ городской среды в рамках Стратегии градостроительного развития города Заполярный',
-        href: '/zapolyarniy',
-        client: 'МЛА+ СПб',
-        partners: '',
-        year: '2020',
-        location: 'Заполярный, Мурманская область',
-    },
-    {
-        name: 'Колористика Норильска. Типовые колористические решения фасадов зданий, строений, сооружений',
-        href: '',
-        client: 'АНО «Агентство развития Норильска»',
-        partners: 'ООО ИТП «Урбаника»',
-        year: '2020',
-        location: 'Норильск, Красноярский край',
-    },
-    {
-        name: 'Обучающий семинар в рамках ежегодного Регонального форума «Городская среда - стратегический вектор развития»',
-        href: '/ugra-edu',
-        client: 'АНО «Центр развития жилищно-коммунального комплекса и энергосбережения Югры»',
-        partners: '',
-        year: '2019',
-        location: 'Ханты-Мансийск, Ханты-Мансийский автономный округ - Югра',
-    },
-    {
-        name: 'Методические рекомендации по подготовке Конкурсных заявок МГиИП',
-        href: '/application-manual',
-        client: 'АНО «Центр развития жилищно-коммунального комплекса и энергосбережения Югры»',
-        partners: '',
-        year: '2019',
-        location: 'Ханты-Мансийский автономный округ - Югра',
-    },
-    {
-        name: 'Второе место в конкурсе на мастерплан села Оймякон - полюс холода',
-        href: '/oymyakon',
-        client: 'Агентство стратегического развития «ЦЕНТР»',
-        partners: 'МЛА+ СПб, ТС-Центр, ПРОМКОД',
-        year: '2019',
-        location: 'Оймякон, Республика Саха - Якутия',
-    },
-    {
-        name: 'Архитектурно-градостроительная концепция IT школы',
-        href: '/yoshkola',
-        client: 'iSpring',
-        partners: '',
-        year: '2019',
-        location: 'Йошкар-Ола, Республика Марий Эл',
-    },
-    {
-        name: 'Концепция обучающего пространства и соучаствующее проектирование',
-        href: '/epischool',
-        client: 'Эпишкола',
-        partners: '',
-        year: '2019',
-        location: 'Санкт-Петербург',
-    },
-    {
-        name: 'Концепция открытых городских пространств и дизайн-код жилого района Скандинавия',
-        href: '/scnd-gb',
-        client: 'ГК «А101»',
-        partners: 'МЛА+ СПб',
-        year: '2018',
-        location: 'Коммунарка, Москва',
-    },
-]
-
-const tableDataGs = [
-    {
-        name: 'Система навигации Елагина острова (ЦПКиО им. Кирова)',
-        href: '',
-        client: 'СПб ГБУК МДК ЦПКиО им. С.М. Кирова',
-        partners: 'ООО «РСК АНТИК»',
-        year: '2024',
-        location: 'Санкт-Петербург',
-    },
-    {
-        name: 'Дизайн-проект и Проектная документация Завода ЕАС Дзержинский',
-        href: '',
-        client: '',
-        partners: '',
-        year: '2024',
-        location: 'Дзержинский, Московская область',
-    },
-    {
-        name: 'Детализированная концепция и РД благоустройства 1-й очереди строительства жилого комплекса "Квартал Торики"',
-        href: 'https://samolet.ru/spb/project/kvartal-toriki/',
-        client: 'ГК «Самолет»',
-        partners: '',
-        year: '2024',
-        location: 'Горелово, Ленинградская область',
-    },
-    {
-        name: 'Конкурсная заявка МГиИП на благоустройство сквера Молодежный',
-        href: '',
-        client: 'Администрация Шалинского района',
-        partners: 'ООО ИТП «Урбаника»',
-        year: '2024',
-        location: 'Шали, Чеченская республика',
-    },
-    {
-        name: 'Детализированная концепция благоустройства 1-й очереди строительства жилого комплекса "Сердце Сахалина"',
-        href: 'https://samolet.ru/korsakov/project/serdce-sahalina/',
-        client: 'ГК «Самолет»',
-        partners: '',
-        year: '2024',
-        location: 'Корсаков, Сахалинская область',
-    },
-    {
-        name: 'Рамочная концепция жилого района "Донские Легенды" и детализированные концепции благоустройства 1-й очереди строительства',
-        href: 'https://samolet.ru/rostov-na-donu/project/donskie-legendy/',
-        client: 'ГК «Самолет»',
-        partners: 'Консультанты: ООО Лабград, ОрганикПанк',
-        year: '2024',
-        location: 'Ростов, Ростовская область',
-    },
-    {
-        name: 'Конкурсная заявка МГиИП на благоустройство улицы Ленина',
-        href: '/verhnyayatura',
-        client: 'Администрация городского округа Верхняя Тура',
-        partners: '',
-        year: '2023',
-        location: 'Верхняя Тура, Свердловская область',
-    },
-    {
-        name: 'Конкурсная заявка МГиИП на благоустройство первого Гостевого прогулочного маршрута',
-        href: '/verhoturie',
-        client: 'Администрация городского округа Верхотурье',
-        partners: '',
-        year: '2023',
-        location: 'Верхотурье, Свердловская область',
-    },
-    {
-        name: 'Конкурсная заявка МГиИП на благоустройство Первого городского парка',
-        href: '/nefteyugansk',
-        client: 'Департамент градостроительства и земельных отношений Администрации г.Нефтеюганска',
-        partners: '',
-        year: '2023',
-        location: 'Нефтеюганск, Ханты - Мансийский автономный округ - Югра',
-    },
-    {
-        name: 'Конкурсная заявка МГиИП на благоустройство улицы Московская',
-        href: '/volokolamsk-2',
-        client: 'АНО «Лаборатория развития территорий городской среды»',
-        partners: 'ООО ИТП «Урбаника»',
-        year: '2022',
-        location: 'Волоколамск, Московская область',
-    },
-    {
-        name: 'Экспертиза и соучаствующее проектирование Бульвара Карла Маркса',
-        href: '',
-        client: 'Департамент градостроительства и архитектуры',
-        partners: '',
-        year: '2022',
-        location: 'Воронеж, Воронежская область',
-    },
-    {
-        name: 'Концепция, ПСД и РД по благоустройству сквера Чекистов',
-        href: '',
-        client: 'Бюро «Плотинка»',
-        partners: '',
-        year: '2022',
-        location: 'Пермь, Пермский край',
-    },
-    {
-        name: 'Шуваловский карьер',
-        href: '/shvl-spb',
-        client: 'Комитет по благоустройству Санкт - Петербурга, Центр компетенций по вопросам формирования комфортной городской среды в Санкт - Петербурге и администрация Приморского района',
-        partners: 'ОрганикПанк',
-        year: '2022',
-        location: 'Санкт - Петербург, Санкт - Петербург',
-    },
-    {
-        name: 'Эскизный проект пользовательского контура, генеральный план, благоустройство, архитектурные решения и интерьеры Завода ЕАС Нарофоминск',
-        href: '',
-        client: 'ООО "Коттех"',
-        partners: '',
-        year: '2022',
-        location: 'Нарофоминск, Московская область',
-    },
-    {
-        name: 'Благоустройство набережной реки Песчанка МГиИП',
-        href: '',
-        client: 'Администрация Соль - Илецкого городского округа',
-        partners: '',
-        year: '2022',
-        location: 'Соль - Илецк, Оренбургская область',
-    },
-    {
-        name: 'Конкурсная заявка МГиИП на благоустройство центральной площади',
-        href: '',
-        client: 'Ярус',
-        partners: '',
-        year: '2021',
-        location: 'Кемь, Республика Карелия',
-    },
-    {
-        name: 'Концепция и РД видовой площадки "Уши Шмидта" в рамках проекта мойзалив.рф',
-        href: '/mur-shmidta',
-        client: 'АНО «Центр городского развития Мурманской области»',
-        partners: 'ООО Цайтплан, ООО knappt, ГНЁМ ВСЁ',
-        year: '2021',
-        location: 'Мурманск, Мурманская область',
-    },
-    {
-        name: 'Дизайн - проект, ПСД и авторский надзор по благоустройству береговой линии и Центрального городского пляжа',
-        href: '/derbent',
-        client: 'Благотворительный фонд “Я люблю Дербент”',
-        partners: '',
-        year: '2021',
-        location: 'Дербент, Республика Дагестан',
-    },
-    {
-        name: 'Благоустройство центральных улиц Перми',
-        href: '/perm',
-        client: 'ГБУ «Институт территориального планирования» Пермского Края',
-        partners: 'МЛА + СПб, OTS lab',
-        year: '2021',
-        location: 'Пермь, Пермский край',
-    },
-    {
-        name: 'Конкурсная заявка МГиИП на благоустройство набережной Верхетуринского водохрнилища',
-        href: '/v-tura',
-        client: 'Администрация городского округа Верхняя Тура',
-        partners: '',
-        year: '2021',
-        location: 'Верхняя Тура, Свердловская область',
-    },
-    {
-        name: 'Концепция благоустройства территории туристическо - рекреационного комплекса Skysoul',
-        href: '/skysoul',
-        client: 'ООО «СЗ «ДК» СКГ»',
-        partners: '',
-        year: '2021',
-        location: 'Коктебель, Республика Крым',
-    },
-    {
-        name: 'Технический отчет для конкурсного жюри в рамках конкурса "5 Углов"',
-        href: '',
-        client: 'АНО "Центр городского развития Мурманской области"',
-        partners: '',
-        year: '2021',
-        location: 'Мурманск, Мурманская область',
-    },
-    {
-        name: 'Комплексный анализ территории и концецпия набережной',
-        href: '',
-        client: 'АНО Центр компетенций по вопросам городской среды Ханты - Мансийского автономного округа - Югры',
-        partners: '',
-        year: '2021',
-        location: 'Радужный, Ханты - Мансийский автономный округ - Югра',
-    },
-    {
-        name: 'Ревизии и технический отчет проектно - сметной документации парка',
-        href: '',
-        client: 'АНО Центр компетенций по вопросам городской среды Ханты - Мансийского автономного округа - Югры',
-        partners: '',
-        year: '2021',
-        location: 'Нягань, Ханты - Мансийский автономный округ - Югра',
-    },
-    {
-        name: 'Эспертиза и графическое оформление Конкурсных заявок МГиИП',
-        href: '',
-        client: 'АНО «Центр развития жилищно - коммунального комплекса и энергосбережения Югры»',
-        partners: '',
-        year: '2020',
-        location: 'Ханты - Мансийский автономный округ - Югра',
-    },
-    {
-        name: 'Конкурсная заявка МГиИП на благоустройство улицы Советская',
-        href: '/volokolamsk',
-        client: 'АНО «Лаборатория развития территорий городской среды»',
-        partners: 'ООО ИТП «Урбаника»',
-        year: '2020',
-        location: 'Волоколамск, Московская область',
-    },
-    {
-        name: 'Конкурсная заявка МГиИП на благоустройство улицы Ленина',
-        href: '/pitkyaranta',
-        client: 'ОАО «Целлюлозный завод «Питкяранта»',
-        partners: 'ООО ИТП «Урбаника»',
-        year: '2020',
-        location: 'Питкяранта, Республика Карелия',
-    },
-    {
-        name: 'Проектная сметная документация на благоустройство улицы Ленина',
-        href: '',
-        client: 'ОАО «Целлюлозный завод «Питкяранта»',
-        partners: '',
-        year: '2020',
-        location: 'Питкяранта, Республика Карелия',
-    },
-    {
-        name: 'Дизайн - проект и Проектная документация по благоустройству улица Канделаки в рамках реализации мастер - плана города Дербент',
-        client: 'УЖКХ города Дербент',
-        year: '2020',
-        location: 'Дербент, Республика Дагестан',
-        href: '/kandelaki',
-    },
-    {
-        name: 'Благоустройство бульвара Рокоссовского в рамках программы Среда НН800',
-        client: 'Институт развития городской среды Нижегородской области',
-        partners: 'Проектная группа «8 линий»',
-        year: '2020',
-        location: 'Нижний Новгород, Нижегородская область',
-        href: '/sreda800',
-    },
-    {
-        name: 'Конкурсная заявка МГиИП на благоустройство парка "Сад Технический поселок"',
-        client: 'Фонд «Наш город»',
-        year: '2020',
-        location: 'Краснокамск, Пермский край',
-        href: '/krasnokamsk',
-    },
-    {
-        name: 'Конкурсная заявка МГиИП на благоустройство парка Центральный лес культуры и отдыха',
-        client: 'АНО Центр компетенций по вопросам городской среды Ханты - Мансийского автономного округа - Югры',
-        year: '2020',
-        location: 'Нягань, Ханты - Мансийский автономный округ - Югра',
-        href: '/nyagan',
-    },
-    {
-        name: 'Конкурсная заявка МГиИП на благоустройство набережной реки Урай',
-        client: 'МКУ «Управление градостроительства, землепользования и природопользования города Урай»',
-        year: '2020',
-        location: 'Урай, Ханты - Мансийский автономный округ - Югра',
-        href: '/bereguray',
-    },
-    {
-        name: 'Предпроетный анализ для комплексной системы навигации территории и парка МЕГА Казань',
-        href: '',
-        client: 'ООО «Мегастрой»',
-        partners: '',
-        year: '2020',
-        location: 'Казань, Республика Татарстан',
-    },
-    {
-        name: 'Конкурс на Парк Дружбы',
-        href: '/gyumri',
-        client: 'Благотворительный фонд IDeA',
-        partners: 'Времяплан, Рубен Мовсисян',
-        year: '2020',
-        location: 'Гюмри, Армения',
-    },
-    {
-        name: 'Эскизный проект благоустройства территории жилого комплекса Homecity',
-        href: '/delta',
-        client: 'PPF Real Estate Russia',
-        partners: 'МЛА + СПб, АМ Сергей Киселев и партнеры',
-        year: '2019',
-        location: 'Москва, Москва',
-    },
-    {
-        name: 'Обучающий семинар в рамках ежегодного Регонального форума «Городская среда - стратегический вектор развития»',
-        href: '/ugra-edu',
-        client: 'АНО «Центр развития жилищно - коммунального комплекса и энергосбережения Югры»',
-        partners: 'АНО «Центр развития жилищно - коммунального комплекса и энергосбережения Югры»',
-        year: '2019',
-        location: 'Ханты - Мансийск, Ханты - Мансийский автономный округ - Югра',
-    },
-    {
-        name: 'Эспертиза и графическое оформление Конкурсных заявок МГиИП',
-        href: '',
-        client: 'АНО «Центр развития жилищно - коммунального комплекса и энергосбережения Югры»',
-        partners: '',
-        year: '2019',
-        location: 'Ханты - Мансийский автономный округ - Югра',
-    },
-    {
-        name: 'Предпроектное исследование открытых городских пространств',
-        href: '/uray-ppi',
-        client: 'МКУ «Управление градостроительства, землепользования и природопользования города Урай»',
-        partners: '',
-        year: '2019',
-        location: 'Урай, Ханты - Мансийский автономный округ - Югра',
-    },
-    {
-        name: 'Концепции благоустройства city4people.ru',
-        href: '/gorprojects-spb',
-        client: 'Фонд содействия развитию городов «Городские проекты Ильи Варламова и Максима Каца»',
-        partners: '',
-        year: '2019',
-        location: 'Санкт - Петербург, Санкт - Петербург',
-    },
-    {
-        name: 'SISU Сервис оптимизации рабочего процесса проектирования за счет автоматизации производства чертежей и ТЭПов',
-        href: '/sisu',
-        client: '',
-        partners: '',
-        year: '2019',
-        location: '',
-    },
-    {
-        name: 'Концепция районного центра на набережной реки Карповка',
-        href: '/kemb',
-        client: 'Администрация Петроградского района Санкт - Петербурга',
-        partners: 'Исследовательская группа «Друзья Карповки»',
-        year: '2018',
-        location: 'Санкт - Петербург, Санкт - Петербург',
-    },
-    {
-        name: 'Конкурсная заявка МГиИП на благоустройство бульвара Карла Марска',
-        href: '/chistopol2',
-        client: 'Фонд "ИРГРТ"',
-        partners: 'МЛА + СПб',
-        year: '2018',
-        location: 'Чистополь, Республика Татарстан',
-    },
-    {
-        name: 'Концепция открытых городских пространств и дизайн - код жилого района Скандинавия',
-        href: '/scnd-gb',
-        client: 'ГК «А101»',
-        partners: 'МЛА + СПб',
-        year: '2018',
-        location: 'Коммунарка, Москва',
-    },
-    {
-        name: 'Первое место в закрытом международном архитектурном конкурсе на концепцию парка и жилого района Скандинавия',
-        href: '/trollgardens',
-        client: 'ГК «А101»',
-        partners: 'МЛА + СПб',
-        year: '2017',
-        location: 'Коммунарка, Москва',
-    },
-    {
-        name: 'Концепция, ПСД и авторский надзор по реконструкции Бульвара Карла Марска',
-        href: '/chistopol',
-        client: 'Фонд "ИРГРТ"',
-        partners: 'МЛА + СПб',
-        year: '2017',
-        location: 'Чистополь, Республика Татарстан',
-    },
-    {
-        name: 'Разработка и производство серия уличной мебели(изделий) для Центрального бульвара',
-        href: '/model4',
-        client: 'Фонд "ИРГРТ"',
-        partners: 'OOO ДивоМеталл, ООО knappt',
-        year: '2017',
-        location: 'Чистополь, Республика Татарстан',
-    },
-]
-
-const tableDataCs = [
-    {
-        name: 'Второе место в конкурсе на мастерплан Экополиса и разработка WEB-3д модели нового города',
-        client: 'Агентство стратегического развития «ЦЕНТР»',
-        year: '2021',
-        location: 'Корсаков, Сахалинская область',
-        href: '',
-    },
-    {
-        name: 'Фотопоток - Инструмент анализа цифрового следа',
-        client: '',
-        year: '2021',
-        location: '',
-        href: '/photostream',
-    },
-    {
-        name: 'Алгоритмический анализ городской среды в рамках Стратегии градостроительного развития города Заполярный',
-        client: 'МЛА+ СПб',
-        year: '2020',
-        location: 'Заполярный, Мурманская область',
-        href: '/zapolyarniy',
-    },
-    {
-        name: 'Предпроетный анализ для комплексной системы навигации территории и парка МЕГА Казань',
-        client: '',
-        year: '2020',
-        location: 'Казань, Республика Татарстан',
-        href: '',
-    },
-    {
-        name: 'mesto.io Платформа цифрового диалога. Облачная геоинформационная система, инструмент картирования городских данных и анализа городской среды',
-        client: '',
-        year: '2020',
-        location: '',
-        href: '/mesto',
-    },
-    {
-        name: 'SISU Сервис оптимизации рабочего процесса проектирования за счет автоматизации производства чертежей и ТЭПов',
-        client: '',
-        year: '2019',
-        location: '',
-        href: '/sisu',
-    },
-]
 
 const Page: React.FC = () => {
     const vertOffset = '6rem'
@@ -613,66 +36,13 @@ const Page: React.FC = () => {
                 position: 'relative',
             }}>
             </div>
-            <WideBlock>
-                <Carousel>
-                    <Image src="/static/1.jpg" width={1920} height={1080} alt="" />
-                    <Image src="/static/2.jpg" width={1920} height={1080} alt="" />
-                    <Image src="/static/3.jpg" width={1920} height={1080} alt="" />
-                </Carousel>
-            </WideBlock>
+            <Carousel>
+                <Image src="/static/1.jpg" width={1920} height={1080} alt="" />
+                <Image src="/static/2.jpg" width={1920} height={1080} alt="" />
+                <Image src="/static/3.jpg" width={1920} height={1080} alt="" />
+            </Carousel>
 
-            <div style={{
-                position: 'relative',
-                scrollBehavior: 'smooth',
-                // paddingTop: vertOffset,
-            }}>
-                <TableButtons />
-
-                <div id='masterplans' />
-                <Title as='h2'>
-                    <>
-                        мастер-планы
-                        <div style={{
-                            fontSize: '1rem',
-                            lineHeight: '1.5',
-                        }}>
-                            дизайн-код / стратегические документы и рамочные концепции / комплексное развитие территорий / архитектурно-градостроительные концепции / цифровой двойник / WEB 3д-модель
-                        </div>
-                    </>
-                </Title>
-
-                <Table data={tableDataMp} />
-
-                <div id='cityspace' />
-                <Title as='h2'>
-                    <>
-                        городская среда
-                        <div style={{
-                            fontSize: '1rem',
-                            lineHeight: '1.5',
-                        }}>
-                            благоустройство территорий: концепция, ЭП, ПСД, РД, авторский надзор / проекты ФКГС / конкурсные заявки МГиИП / разработка уникальных объектов и МАФ / соучастие и социокультурное программирование / экспертиза средовых проектов
-                        </div>
-                    </>
-                </Title>
-
-                <Table data={tableDataGs} />
-
-                <div id='digitalspace' />
-                <Title as='h2'>
-                    <>
-                        цифровая среда
-                        <div style={{
-                            fontSize: '1rem',
-                            lineHeight: '1.5',
-                        }}>
-                            цифровой двойник / WEB 3д-модель / инструменты городских исследований / анализ цифрового следа / онлайн платформа сбора мнений
-                        </div>
-                    </>
-                </Title>
-
-                <Table data={tableDataCs} />
-            </div>
+            <IndexTables />
 
             <div id='team' style={{
                 position: 'relative',
