+++ src/components/GalleryPartners/item.module.css
@@ -3,14 +3,14 @@
     display: flex;
     flex-flow: column;
     overflow: hidden;
-    /* min-height: calc(var(--cell-size-min) + var(--grid-gap)); */
 
-    cursor: pointer;
     text-decoration: none;
 
-    border: solid 1px var(--color-background-opposite);
-    transition: all 0.5s ease;
+    /* border: solid 1px var(--color-background-opposite); */
+    /* transition: all 0.5s ease; */
+    filter: grayscale(1);
 }
+
 .item:hover {
     border-color: var(--link-color-active);
 }
