+++ src/components/Index/indexTables.module.css
@@ -1,0 +1,38 @@
+.tables {
+    position: relative;
+    scroll-behavior: smooth;
+    padding-top: 3rem;
+}
+
+.container {
+    position: sticky;
+    top: 3.75rem;
+    background: white;
+    z-index: 2;
+    padding: 1rem 0;
+}
+
+.container a {
+    text-align: center;
+    text-decoration: none;
+    font-weight: bold;
+    font-size: 1.5rem;
+    background-color: var(--color-background-opposite);
+    color: var(--color-text-opposite);
+    padding: 0.5rem 1rem;
+}
+
+.container>div {
+    margin: 0;
+}
+
+@media screen and (max-width: 960px) {
+    .container {
+        display: none;
+    }
+
+    .tables h2 {
+        font-size: 1rem;
+        line-height: 1.5;
+    }
+}
