--- src/components/Index/tableButtons.module.css
@@ -1,18 +1,0 @@
-.container {
-    position: sticky;
-    top: 3.75rem;
-    background: white;
-    z-index: 1;
-    padding: 1rem 0;
-}
-
-.container>div {
-    margin: 0;
-}
-
-@media screen and (max-width: 960px) {
-    .container {
-        position: relative;
-        top: 0;
-    }
-}
